requirements.txt
networkx
pandas
geopy
staticmap
haversine
python-telegram-bot==10
